Moved `idesync` tests to `test_tools.py`; improved `idesync` output; docstring edits
