+++ testing/test_sync.py
@@ -5,7 +5,6 @@ import shutil
 import pytest
 
 from idelib import importer, sync, userdata
-from idelib.tools import idesync
 
 
 SYNC_INFO = {
@@ -469,41 +468,3 @@ def test_gnss_sync_userdata(tmp_path):
         assert 'TimeBaseUTCFine' in doc._userdata
         assert doc.currentSession.utcStartTime != doc.currentSession.utcStartTimeOriginal
         assert doc.currentSession.utcStartTime == gnsstime
-
-
-# ===========================================================================
-# Utility tests
-# ===========================================================================
-
-def test_idesync_basics(tmp_path):
-    filenames = []
-    cwd = os.path.dirname(__file__)
-    for f in ('GNSS1.IDE', 'TSF1.IDE', 'TSF2.IDE'):
-        shutil.copy2(os.path.join(cwd, f), tmp_path / f)
-        filenames.append(str(tmp_path / f))  # Filenames will be strings when run from the CLI
-
-    # One file: just set GPS/GNSS time base
-    idesync.syncFiles(filenames[0], gps=True)
-    with importer.importFile(filenames[0]) as doc:
-        userdata.readUserData(doc)
-        assert doc._userdata['TimeBaseUTCFine'] != doc.currentSession.utcStartTimeOriginal
-
-
-def test_idesync_fail(tmp_path):
-    filenames = []
-    cwd = os.path.dirname(__file__)
-    for f in ('GNSS1.IDE', 'TSF1.IDE', 'TSF2.IDE', 'test3.IDE'):
-        shutil.copy2(os.path.join(cwd, f), tmp_path / f)
-        filenames.append(str(tmp_path / f))  # Filenames will be strings when run from the CLI
-
-    # Fail: can't use `gps` and `inherit` together
-    with pytest.raises(sync.SyncError, match=r'.*mutually exclusive.*'):
-        idesync.syncFiles(filenames[1], filenames[2], gps=True, inherit=True)
-
-    # Fail: No GPS/GNSS time data
-    with pytest.raises(sync.SyncError, match=r'No GPS/GNSS data.*'):
-        idesync.syncFiles(filenames[1], filenames[2], gps=True, inherit=False)
-
-    successes, failures = idesync.syncFiles(filenames[0], filenames[1], filenames[2], gps=False)
-    assert len(successes) == 0
-    assert len(failures) == 2
